Conflicts:
	Bermuda/Bermuda/Game.cpp
	Bermuda/Bermuda/ImageLoader.h
	Bermuda/Bermuda/PlayState.cpp

+++ Bermuda/Bermuda/Game.cpp
@@ -47,7 +47,7 @@ void Game::gameLoop(GameStateManager* gsm) {
 		//Take current Time
 		QueryPerformanceCounter(&currentTime);
 		//Calculate difference (previousTime - currentTime)
-		long updateLength = (currentTime.QuadPart - previousTime.QuadPart) * 1000.0 / frequency.QuadPart;
+		long updateLength = static_cast<long>((currentTime.QuadPart - previousTime.QuadPart) * 1000.0 / frequency.QuadPart);
 		
 		//Previous time = current time
 		previousTime = currentTime;
@@ -82,7 +82,7 @@ void Game::gameLoop(GameStateManager* gsm) {
 		QueryPerformanceCounter(&afterLoopTime);
 		//Sleep if needed (if time took longer than optimal)
 		if ((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME) > 0) {
-			Sleep((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME));
+			Sleep(static_cast<DWORD>((((previousTime.QuadPart - afterLoopTime.QuadPart) * 1000.0 / frequency.QuadPart) + OPTIMAL_TIME)));
 		}
 	}
 }
